getParcelableArray* behaviour is sometimes unpredictable
in some case when the data in a savedInstance the Parcelable arrays stored by the application are not unparcel'ed correctly because the ClassLoader doesn't seem to be correct.
The first getSparseParcelableArray() in the code is preceded by setClassLoader() by not the second one. The problem disappears when setClassLoader() is called in both cases.

Change-Id: I2d2f42c285d9130fd543c0154e9d210d430823e3