Update patch set 1
Patch Set 1:

The correct Commit Message:

In some case when the data in a savedInstance Bundle the Parcelable arrays stored by the application are not unparcel'ed correctly because the ClassLoader doesn't seem to be correct (doesn't have the classes from the app).

The second getSparseParcelableArray() in the code is preceded by setClassLoader() but not the first one. The problem disappears when setClassLoader() is called in both cases.

Patch-set: 1
Reviewer: Steve Lhomme <1014910@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0