Update patch set 1
Patch Set 1: Do not submit

Could you please elaborate on the bug this aims to solve, or provide a repro case?

onLayout should always produce a layout consistent with the current state. Skipping a step of the layout while scrolling does not appear to be the right thing here; t a glance this seems to introduce buggy behavior. If there is a bug in updateTextPositions causing an issue, we should fix that.

Patch-set: 1
Label: Code-Review=-2