Update patch set 1
Patch Set 1:

If the bug also occurs in AOSP, a more correct solution might be to take into account the offset when drawing the labels so that the correct ones get drawn.

I looked into doing that at first and could not find a way to determine if the view had been scrolled to the left or the right in order to know which labels needed to be drawn.

Patch-set: 1