Update patch set 1
Patch Set 1:

I tried this exact change before creating the other one I had submitted.

This resolves the issue when scrolling to the right, but actually causes more problems when scrolling left.  If you scroll left the offset is still between 0.0f and 1.0f.

If you scroll slightly to the left (less than 50% of the way to the next item) the offset will be >0.5f and it will increment the position by 1. In this situation it should not be changing the position at all because you're still over the same item.

If you scroll all the way to the left (more than 50% of the way to the next item) the offset will be <0.5f and the position will not be changed when in fact it should be decremented by 1.

Patch-set: 1