Update patch set 1
Patch Set 1:

Prior to Jelly Bean, Android had a bug in handling 'homeAsUp' as part of the display options in an action bar style.  It would create the up indicator, but would not enable the application home button.  The fix for this can be seen at:
  https://android.googlesource.com/platform/frameworks/base/+/d40423a7adff41463fb4a563becc0bd77fe9919c
I believe the ActionBar provided by the support library should account for this, and give consistent behavior.
That is what my aim is with this patch.

Patch-set: 1
Reviewer: Jeffrey Smith <1022047@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0