Fix SlidingPaneLayout padding
When measuring the children width take into account the layout
horizontal padding.

When reporting the layout measured height, take into account the
vertical padding that was removed to measure the children.
Also make sure the width and height respects the minimun suggested.

And when sliding the pane, make sure views are not offset vertically.

Change-Id: If17912b94fa6721c199ad1a050d8c7350738076b
Signed-off-by: Sérgio Faria <sergio91pt@gmail.com>