Fixed reordering crash with FragmentPagerAdapter.
Made ViewPager's dataSetChanged() call the new PagerAdapter method moveItem()
for all items that change position. FragmentPagerAdapter implements this and
removes the Fragment in order to re-add it with the updated tag. This ensures
subsequent calls to populate() no longer result in IllegalStateExceptions with
regard to changing the Fragments tag.

Also inadvertedly addresses issue:
https://code.google.com/p/android/issues/detail?id=42601

Change-Id: Ic826186139f1f46562e432ed5c3b9033c63f62a5
Signed-off-by: Benjamin Dobell <benjamin.dobell+git@glassechidna.com.au>