Fixed FragmentPagerAdapter memory leak.
FragmentPagerAdapter adds Fragments to the FragmentManager when instantiating an item, but previously never removed them when the data set changed. FragmentPagerAdapter was detaching them (i.e. the FragmentManager still had a reference), which is incorrect for an item removed from the data set, but even then this was only occurring for Fragments that had their View in memory at the time the data set changed.

FragmentPagerAdapter now keeps track of the fragments it adds to the fragment manager and removes old fragments when notifyDataSetChanged() is called.

Change-Id: I6c0e3eb7e133bdf83249d0b7823ca2050a3e7692
Signed-off-by: Benjamin Dobell <benjamin.dobell+git@glassechidna.com.au>