Update patch set 1
Patch Set 1:

That import is correct. The Makefile change looks good to me, though Xav is the expert on how the support library is built.  (We didn't use to need to include the support library explicitly in libraries that depend on v4-support; it would come transitively from the support dependency, but maybe that's changed recently?)

Patch-set: 1
Reviewer: Tor Norbye <1004689@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0