Update patch set 1
Patch Set 1: Verified-1

This is an automated message from the Deckard Autoverifier.

Patchset merges but does not build in AOSP.
The following snippet of `make` output may point you in the right direction:

  make: *** No rule to make target `frameworks/rs/driver/rsdElement.cpp', needed by `out/target/product/generic/obj/SHARED_LIBRARIES/libRSSupport_intermediates/driver/rsdElement.o'.  Stop.



If your patchset has a dependency in a separate project, please ensure that
both changes are uploaded with the same Change-Id.

Patch-set: 1