Update patch set 1
Patch Set 1:

The framework implementation uses and has always has used setSaveFromParentEnabled(false).

This patch changes the View given to onViewCreated, so it may or may not be a wrapper depending on the platform version, instead of always being a wrapper. With this change, new code that attempts to cast that View will only get a ClassCastException on older platform versions, instead of always.

This potential headache can be solved, perhaps, with a new lint check and better documentation in onViewCreated, that right now wrongfully says the view is the same returned in onCreateView. It is the same on the framework implementation, but always a FrameLayout wrapper in the support package...

Patch-set: 1
Reviewer: Sérgio Faria <1020767@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0