Fix memory leak when fast scrolling rows.
If a presenter starts a view property animation, then the parent row view
will have transient state which may cause a row to fail to be recycled
during a fast scroll, because RecyclerView checks transient state and
refuses to recycle a view if it or its children has running view property
animation.

This can cause a memory leak because ObjectAdapters have references to
ItemBridgeAdapters via the registered observer mechanism.

Apps should clear any view property animations in Presenter
onViewDetachedFromWindow, but in case they don't we'll do it for them in
the base class.

b/17013302

Change-Id: Ibdf5998e81dd130128f88f85d88243ec27a70dd5