Update patch set 2
Patch Set 2:

Can we get any movement on this?  This is affecting downstream tools that use Maven for resolving the dependencies.  The poms are not valid maven pom files and this patch corrects that so that it not only works with Gradel but also maven.

Patch-set: 2
Reviewer: David Carver <1043503@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0