Update patch set 1
Patch Set 1:

Thank you for your suggestion!
I'm thinking about using "own round rule" and add extra pixel not if fraction part of size from the start of the row is 0.5 and bigger, but if ((1 - currentFraction) < mSpanSizeFraction) (maybe, it's better to use integers for better perfomance and accuracy). in your example (1 - 0.8) < 0.4.
And i need to know view's position in a row (in terms of spans).
As i understood, if layout in primary direction (line 339), layoutChunk will work for whoole row. 
Otherwise, ... i will know tomorrow, because i'm in +3 UTC and it's 10PM. And i'm asking. not fully understood, because i want to know, if i'm moving in the right direction.

Patch-set: 1