Update patch set 1
Patch Set 1:

layoutChunk will always be called for a whole row. (assuming list if vertical).

If you want to implement, here is my suggestion:

have an int[] with size mSpanCount + 1. Lets call this mKeylines array.

mKeylines[0] is always getStartAfterPadding.
mKeylines[mSpanCount] is getEndAfterPadding.
for the positions in between, you distribute it as equal as possible.

In layout chunk, you'll first decide between which keylines an item will be laid out. 
Then the space that will be given to it will be between these keyline.

Assume we have 5 spans. (indexed from 0 to 4).

If a view view occupy spans 2,3; it's MeasureSpec will be mKeylines[4] - mKeylines[2].

Lmk if this is getting confusing, then we'll implement it internally.

Thanks.

Patch-set: 1