Added option for custom View Titles
This extension allows for custom views to be set in the implementing class.  To take advantage of this, the custom view should be checked if it exists before checking for a title.  If the custom view call is null, as in the default implementation, the caller should continue to check the title before resulting to a default value to identify the page.