Fixed ViewPager math for PageTransformer with paddingLeft
Since the ViewPager calculates transform positions based off of each child's left position the ViewPager needs to account for the fact that the left position is increased when there is left padding. This causes PageTransformers to be called with invalid transform positions whenever left padding is used on a view pager, making it difficult to write a usefull PageTrasnformer when left padding is present. Right padding is already accounted for because the math is based on the clientWidth wich accounts for padding because the view is sized to fit within the padding. The fix for this was to update the math for the transformPos to (child.getLeft() - scrollX - paddingLeft) / getClientWidth()