Read clip to padding from attrs
GetClipToPadding is API 21 so RecyclerView steals that
vlaue by overriding setClipToPadding and tracking the
value itself.

Unfortunately, when clipToPadding is defined in the XML and set
by the parent's constructor, RV's field initialization was overriding
the value.
Apperantly, the R.attr.clipToPadding is from API 1
so we can use it to read the value (getter is from 21).

Bug: 29123702
Bug: 26779229
Change-Id: I0521413bbf81c6c5428e7305eefa838c988fc018