Fail focus search if focused view disappears during focus search
This CL fixes an edge case where the last focused view disappears
during a focusSearch and we need to call focus search failure.

Calling super potentially crashes the app and calling the LayoutManager
invalidates the API. For now, best seems to just fail the focus
search. Later, we can get more clever and try to pick a view
in a position around the previously focused view.

Bug: 27837873
Change-Id: Id8d5e8ea2e3ce40982e0055b0f45777ef8b2b055