Speculative fix for exception in CoordinatorLayout.onLayout
Caused by a prepareChildren call in onChildViewRemoved().
If a child is removed in the middle of a layout pass, the
count is wrong and we hit an IndexOutOfBoundsException.

Fixed by removing the prepareChildren call, which also fixes
onDependentViewRemoved() which was broken in
I1283f72c312824ee3ed1fd4b0bb7e7405f38cb2b.

BUG: 30252077

Change-Id: Ifce3178cf69685d519c859c2ebd6402b744204b1