Fixes IntRange value of Snackbar Duration IntDef
See https://code.google.com/p/android/issues/detail?id=213243

The IntRange value defined for the Duration IntDef in the Snackbar
class was 'from = 1' which did not allow using the
pre-defined duration integer constants like 'LENGTH_INDEFINITE'.
This was solved by setting the IntRange to start at 'LENGTH_INDEFINITE'
which is a value of '-2'. This assumes that 'LENGTH_INDEFINITE' will
always be the lowest integer value that a Snackbar duration can use.

Test: Existing Snackbar instrumentation tests passed

Change-Id: Ide4ff6390051341d3118a32de39d30a6bc0ee628
Signed-off-by: Tal Kirshboim <tal.kirshboim@gmail.com>