Update patch set 1
Patch Set 1:

> Are you trying to override them? If so, it already calls
 > onAnimationFinished.
 > Or is there another reason for this?

Basically, the idea is to make ItemAnimators composable. Imagine you have a library with 2 ItemAnimators that implement some fancy animations, but you want to use the appear animation of the first one and the disappear animation of the second one. Now you try to combine those into a new ItemAnimator which calls the appear animation of ItemAnimator1 and disappear animation of ItemAnimator2 (my specific use case is to delegate the change animation logic to an item specific class similar to what others have already done with the adapter, but it doesn't matter). This is currently not (easily or at all) possible because RecyclerView and LayoutManagers want to be notified when animations end.

I guess one could subclass both LayoutManagers, override onAnimationFinished like you suggested and implement yet another listener, but I find that very messy since it's not generic and the same logic is already implemented, just inaccessible.

To make it truly composable there has to be some way for our composed class to know when each animation in the different ItemAnimator classes ended and in turn notify RecyclerView and/or any LayoutManagers.

Now that I think about it, the same may be more easily achieved by making ItemAnimator.setListener protected/public. That way the composed class can itself set a listener and call dispatchAnimationFinished, without a need to change any of the dispatch* methods. Let me amend that in a minute. It doesn't provide a way to be notified about animation starts though, which would require adding a new method to the listener interface.

Patch-set: 1
Label: Verified=0