Update patch set 2
Patch Set 2:

Unfortunately, composing animators is not as easy.
Animations need to be continious such that if an item is moving then suddenly disappears, ItemAnimator should gracefully do the migration.
If these are two different animators, i cannot see how it works unless they agree on another API.

onAnimationFinished is a method in ItemAnimator so there is no difference between overriding it or dispath.
Of course, if you override dispatch, you can prevent the call from going into the RV but if you are trying to fool the ItemAnimator and run it w/o a RV, you'll have bigger problems.

tl;dr; for what you are trying to do, you need another API contract where this new type of ItemAnimators agree and this contract should handle continiuity. Current design of ItemAnimators are no suitable for that.

Patch-set: 2