Enabled customization of the arrow color in OnboardingFragment
Previosuly, we could customize the foreground color of the arrow but not
the arrow itself which was a PNG drawable. This CL uses ColorFilter that
is installed on the paint which draws the arrow. This enables setting
the bitmap color.
Also defined theme attribute for setting color. Programmatically setting
this color will supersede the theme attribute.

Bug: 37881573
Test: tested with a subclass of OnboardingFragment both setting the
color through the theme attribute as well as dynamically in onCreate().

Change-Id: I18a8d98f7fffe1c029a9116bbcd031ad67c6b10d
(cherry picked from commit 0cb1f577feba17baceb159ab7592e1050425bbad)