Update patch set 3
Patch Set 3:

For the record, I found a way to dodge the bug with the latest support libraries:
https://gist.github.com/cbeyls/ca39c38f1ad451c35fc66142989e9e76

During a config change in FragmentActivity I call mFragments.doLoaderStop(false) manually right after mFragments.doLoaderStop(true) has been called, which has the effect of setting the FragmentHostCallback.mRetainLoaders flag back to false. LoaderManagers that are not started at that point will be destroyed and later re-created instead of forcefully start/stopped in retainLoaderNonConfig(). Additionally I also need to clear the cached mLoaderManager reference in retained fragments to prevent them to use the destroyed instance.

I use this hack in production for apps with thousands of installs and had zero issues so far, but of course it's not pretty and only works with the current implementation.

Patch-set: 3
Reviewer: Christophe Beyls <1139359@85c56323-6fa9-3386-8a01-6480fb634889>