Make NonNull & Nullable a type annotation.
Nullability itself is a "type use" information. This means
it is a modifier on a reference, not a property of a class,
method or a parameter. "Type use" annotations were introduced
in openJdk8 and are available in android O and later.

This chance will make type annotatios coexist with method/parameter
annotations (as expleined in jls 9.7.4) so it won't cause
any compatiblity issues.

Test: Not applicable
Bug: 64930165
Change-Id: Ic41cceea443e55d5394a5481cb9e910c5c32385e