Update patch set 4
Patch Set 4:

No, that doesn't mean the same thing as @AnyThread. It means that the method can be called from any of the explicitly mentioned threads, but not any other ones.

For example, if we have this:

@WorkerThread @BinderThread methodA()
@BinderThread methodB()

Then it is safe for methodB to call methodA (since methodB is always on the binder thread which is one of method A's allowed threads), but not the other way around (methodA may be on the WorkerThread, not the BinderThread, so when it invokes MethodB it is on a thread that is not allowed.)

Patch-set: 4
Reviewer: Tor Norbye <1004689@85c56323-6fa9-3386-8a01-6480fb634889>