Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> This is a breaking API change - you've changed the signature of an existing Behavior method, instead of deprecating it and introducing a new one. Also, if all existing behaviors need to be switched to the new method, that's a rather high barrier to clear for this CL.
> 
> Also note that since the first revision of this CL, CoordinatorLayout has moved out of the design package, and the design library itself is now in a different project - https://github.com/material-components/material-components-android/tree/master/lib/

I've fixed the public API breaking change. Also there is no need for an existing apps to switch to a new method; the old one is still called if the new one is not overrided.

I'll create the PR to the material-components repo as well. But what about this one? This is some outdated code or what version of CoordinatorLayout is in this repo?

Thanks.

Patch-set: 3
Reviewer: Miroslav Kacera <1197023@85c56323-6fa9-3386-8a01-6480fb634889>