SeekBarPreference fixed on API 14-19
SeekBarPreference doesn't have a background set which triggers a
padding bug when setting the background dynamically in
PreferenceGroupAdapter.
This commit fixes that by defining the background attribute on the
root element of all versions of
preference_widget_seekbar_material.xml.
Also, this commit fixes the SeekBar not respecting the
paddingStart and paddingEnd attributes on API 17-19.
The issue was reported as 111405638.
Test: tested the changes using API 15 and API 17 emulators.

Change-Id: Ia7d71ff0d1089a80c23990943e86b18d4dc8a8d3

Additional page:
https://android-review.googlesource.com/c/platform/frameworks/support/+/717787