Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > Hi,
> > 
> > Could you move this change into the other change? Or is there a specific reason why you sent this one separately?
> > 
> > Thanks!
> 
> Hi,
> 
> No, there's no particular reason why this is a separate one. I just didn't know how to add to an already uploaded repo as I discovered only after submitting it that this was left out. Could you provide the necessary commands / guide how to merge this one into that (or create a new commit to that changeset)?
> 
> Thanks!
Gerrit tracks where to upload the change based on Change-Id: in the commit message. So if you make a new commit with both of the changes and just reuse one of the existing Change-Id: in the commit message, then it will override one of your changes you created. Then you can abandon the other change. Does that help?

Patch-set: 3