recyclerview: fix scrolling removes pending moving view.
In most cases we dont need call endAnimation() when scroll pass remove a view
because when view is detached, ViewPropertyAnimation will end. But in some
corner cases, for example if the animation is based on ObjectAnimator or
if the ItemAnimator uses "pending runnable", the ItemAnimation on the view may
not be cleared.

In b/720547, the View is removed by scroll pass while it's in the
pendingMoving list of DefaultItemAnimator and DefaultItemAnimator in a post
runnable perform a postDelayed() on the detached view which never gets executed.

To fix the issue, we issue endAnimation() in scroll pass removing view.
This would make RV more robust, for example in this bug, DefaultItemAnimator
will not have chance to execute pending move animation if the View is removed
in scroll pass.

Bug: 73552923
Test: gradlew recyclerview:connectedCheck -Pandroid.testInstrumentationRunnerArguments.class=androidx.recyclerview.widget.LinearLayoutManagerTest#scrollRemovePendingMovingViewBug73552923
      RecyclerViewAnimationsTest#scrollPassRemoveAdditionView
      RecyclerViewAnimationsTest#scrollPassRemoveMovingView
Test: gradlew leanback:connectedCheck -Pandroid.testInstrumentationRunnerArguments.class=androidx.leanback.widget.GridWidgetTest#testScrollAndRemove
Change-Id: I592a63d17a9b9efbc626b7f0cc4acd1d664bb008