Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> I know we have asked you to combine patches, but when we did I forgot to mention that you should move this to androidx-master-dev branch. That is has all the testing and merging set up. You can move it by using cherry-pick function in gerrit Web UI and typing in androidx-master-dev.
> 
> Thanks for fixing up the follow ups. We'll hopefully straighten out the process soon.

No problem, this is my first actual contribution to the AOSP so it's all new to me. I couldn't decide which branch to use so I picked 1.0 instead of the master (was a bad decision after all lol).
Used cherry-pick so it should be on androidx-master-dev now.

Patch-set: 4
Reviewer: Gerrit User 1084154 <1084154@85c56323-6fa9-3386-8a01-6480fb634889>