Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> Hi,
> 
> Thanks for the change!
> 
> Could you also attach (to the bug you reported 111405638) a before and after screenshot for the padding bug when setting the background dynamically?

Hi,
Added some screenshots to the issue.

Would it be best to include a bug fix for the dynamic background setting too? PreferenceGroupAdapter's onCreateViewHolder(...) looks like this now (this is the source of the bug):

    // ...
    if (view.getBackground() == null) {
        ViewCompat.setBackground(view, background);
    }
    // ...

Which would become this:

    if (view.getBackground() == null) {
        final int[] pad = new int[4];
        pad[0] = ViewCompat.getPaddingStart(view);
        pad[1] = view.getPaddingTop();
        pad[2] = ViewCompat.getPaddingEnd(view);
        pad[3] = view.getPaddingBottom();

        ViewCompat.setBackground(view, background);

        ViewCompat.setPaddingRelative(view, pad[0], pad[1], pad[2], pad[3]);
    }

So it would save the padding, apply the background, the apply the padding again.

Patch-set: 1
Reviewer: Gerrit User 1084154 <1084154@85c56323-6fa9-3386-8a01-6480fb634889>