Fetch Pojo relationships before Pojo creation.
This change lifts the restriction of forcing a setter for a @Relation
annotated field instead of a constructor param. It also fixes the setter
being called early with a non-resolved relationship collection.

To fetch relationships early, this change introduces logic on the
generated code to copy the resulting Cursor of a query into a
MatrixCursor that can be used to iterate more than once. The first
iteration is to retrieve the parent keys of the relations to fetch and
the second is the usual cursor-to-pojo conversion.

Bug: 65772137
Test: ./gradlew :room:room-compiler:test \
      :room:integration-tests:testapp:cC

Change-Id: Ie8043cc6275f67b73f23cca54122e1a37d22d50f