Added clickable span support in pre-o..
In this situation, we capture the relavent info of a clickable span in
the extras bundle, frp, which the service developer can use the support
library getText() to get the spanned text.
An app developer just needs to enable clickableSpan support.
On perform, on the app side, a tag is set with the clickable spans and
ids so it is clear what onClick method to call.

Fix: 72226075
Test: unit tests and gonna verify in sample
app+service.

Change-Id: I2289773f3514a6659eb7a5d7355fffe826e26f42