Always handle onScrolled in ScrollEventAdapter
The logic in ScrollEventAdapter.onScrolled was skipped if there were no
listeners, but the state changes that can occur as a result of
onScrolled are also necessary if there are no listeners.

The regression test that is added to check this behaviour is a bit
complicated, because we need to wait for the scroll to finish, but we
can't listen to it (because we need to test it without listeners). Thus,
we have to poll viewPager.currentItem, but we have to do this on the UI
thread because getCurrentItem can throw a NullPointerException if
setCurrentItem is in progress. Running the poll on the UI thread makes
sure that setCurrentItem is not currently doing anything.

Bug: 109733392
Test: ./gradlew viewpager2:connectedCheck

Change-Id: I1ed0c9d1f0bf30b3d3ddad10b3376f18a9bfaffa