leanback-preference: themerize all colors and fonts
Undo the change of changing default brandColor in leanback, it
should be done in leanback-preference.
Fixed bug that app preference screen built with new library does not
match old platform color. Had a PreferenceThemeOverlayLeanbackBase
to be overlayed differently based on API level.
This is also a first step to have a device based theme strategy so
that single OEM can customize mutliple products with different
settings themes.
Customizable attributes are:
 - settings background color
 - title panel background color
 - title text apperance
 - accent color for category/checkbox/slider/switch
 - item text appearance
 - item second text appearance
 - category text appearance
 - start/end padding
This CL is not trying to theme the layout size.

Bug: 111900425
Test: run same app on both P and <P, see two different sets of colors
and match each platform's settings color.

Change-Id: I20bceba39abe6b8021b99116f004c1e79f957d3d