DO NOT SUBMIT Add ability to detect affected modules by changes
This CL adds a new capability to discover which changes are
made locally. This allows us to avoid copying test apks to
distribution folder, which should improve presubmit times.

By default, it does not do anything. You need to invoke the
build with -Pandroidx.enableAffectedModuleDetection to enable
it.

If a file outside a module is changed (e.g. buildSrc or root),
we re-run all tests.

If no changes are made in frameworks/support, we re-run all
tests.

We also re-run all wear tests for now because their presubmit
hook requires to find it.

Currently, this CL discovers the list of changes by finding
the last Merge commit. We'll eventually move to prop files
when TreeHugger provides us the right information.

Bug: 112058202
Test: GitClientTest, ProjectGraphTest, BuildPropsParserTest
Change-Id: Iebb52b84a5ee025017299e3c6b7e1d77e3e2943d