Remove session releated code
Problem:
Currently, it is not intuitive to developer that what the lifecycle
of a TC should be. TC developer may not be aware that TC is actually
a short-lived session, and there could be a lot of TC around.
From client perspective, it is also difficult to use as a TC can't be
reused once it is destroyed.

Solution:
TextClassifier is no longer session aware by default and thus we can
always reuse the same TC. We will stream events (with TCContet) to the
TC, and request APIs like suggestSelection, classifyText and
generateLinks will now take a TextClassificationContext.
Thus, TC implementation can do whatever they want accordingly,
and define what does a session mean themselves.

1. Removed destroy and isDestoryed
2. Removed SessionStrategy, so no more default session handling.
3. Removed TextClassifierFactory, as we can always reuse the same TC now.

TODO: Another change is needed to have
suggestSelection, classifyText and generateLinks to take
TextClassificationContext.

Bug: 72307474
Bug: 112155387

Test: ./gradlew textclassifier:connectedCheck --info --daemon
Change-Id: Ida3b58d75a809b352e935e32a6517dd29ba7339d