Update patch set 1
Patch Set 1:

My primary concern here is that we'll have to do this for every listener. It'll be a bad experience when these are available for 75% of the types of listeners you are using. When a new version of platform APIs come out we'll have to find all of their listeners and make sure we have copies in this form. There is also all of the listeners in the various AndroidX libraries like ViewPager and AppCompat. Are we going to do all of their listeners?

I've worked on a few libraries that have had this problem (fest-android, assertj-android, RxBinding) and it's not fun to have this constant push/pull of the community wanting APIs for the things they use and you trying to figure out how to not duplicate the entire API surface.

Patch-set: 1
Reviewer: Gerrit User 1232209 <1232209@85c56323-6fa9-3386-8a01-6480fb634889>