Port FloatingToolbar to AndroidX
This CL ports the FloatingToolbar to the Support Library.
The toolbar tries to maintain the style of the toolbar for each
Android version (the look and content of the toolbar has changed
from version to version).

FloatingToolbar.java is basically a copy from the platform with
a few changes to match the use cases we expect to use them in
Support Library (e.g. Smart Linkify).
Changes include:
- Removed the insetsComputer. This is hidden platform API and it
  is used to allow touch events go through the toolbar window so
  users are able to interact with the selection handles that tend
  to be behind (the transparent part of) the toolbar window.
  In Support Library, we don't use the FloatingToolbar where the
  selection handles are visible.
- Clicking outside the Support Library FloatingToolbar always
  dismisses the toolbar. The toolbar is also dismissed when the
  user clicks on the transparent part of the toolbar.

NOTE that we plan to redesign the floating toolbar widget in the
near future so doing a lot of work on it now may not be worthwhile.

Test: Ran androidx.appcompat.widget.FloatingToolbarTest
Change-Id: I58b162924ac3a2d94a28109212e6750d211cbbdf