Update patch set 2
Patch Set 2:

I've pushed metalava 1.0.4 with support for this syntax now. (I've also updated the code which generates the default values a bit; it used to put the source text in there verbatim; that was safe for things like constants/literals, but not for more general expressions; now it visits the AST to make sure that for example symbol references are fully qualified etc. In short, you may want to regenerate the signature files in case they're affected by this.)

Patch-set: 2
Reviewer: Gerrit User 1004689 <1004689@85c56323-6fa9-3386-8a01-6480fb634889>