Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Add copyright and pass pylint and then +2 :)

Oh dear. pylint has a lot (409) of opinions (many of which I don't agree with, or seem not worth the effort to fix).

It'll probably be a long time before I address them, so in the meanwhile I'll remove all of the reviewers from this patch so it doesn't appear in their dashboards until then.

Patch-set: 1
Reviewer: Gerrit User 1137668 <1137668@85c56323-6fa9-3386-8a01-6480fb634889>