Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> Will SliceProvider eventually extend ContentProviderWithCallbacks?

Yes, in a followup I'll integrate with SliceProvider
> 
> Currently all three implementations AppWidgetProvider, BroadcastReceiver and ContentProviderWithCallbacks assume a base class. But in larger projects its often difficult to change the base class or that is abstracted out (eg AGSA :D) 
> 
> Should we also have an API to create a Callbacks objects as a member variabless?

We had discussed this a little bit, but hadn't really landed on anything. I hand't thought about it as a variable though, like a way to tag a member of a broadcast receiver as a callback receiver?

Patch-set: 8
Reviewer: Gerrit User 1042886 <1042886@85c56323-6fa9-3386-8a01-6480fb634889>