Update patch set 7
Patch Set 7: Presubmit-Ready+1

(5 comments)

SimpleGestureGenerator.kt refactored in response to comments.  Ultimately, I really shouldn't be writing another test util for the creation and dispatching of MotionEvents to a View as one already exists in CTS, but I hope that for this CL, it's ok that I have what I have so far.  I will start another CL, right now, to begin the copying and porting over the tool from CTS and then will have a 3rd to move the tests in NestedScrollView and RecyclerView over to use it.

That ok?

Patch-set: 7
Reviewer: Gerrit User 1344416 <1344416@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1