[Work in Progress] Public SmartLinkify APIs
The idea is:
If the API takes TextView, we use AsyncTask as the spannable needs
to be handled in the main thread.

If the API takes Spannable, we just execute everything in
the  calling thread synchronously. Then the app can set the spannable
to TextView or other widgets. This can be seens as the drop-in replacement
of request.genereateLinks(...).apply(...)

Change-Id: Ibff813de35e826190ea29ef1bbe4b26b122a1fbb