Update patch set 8
Patch Set 8:

> Patch Set 8: Code-Review+1
> 
> > Patch Set 8:
> > 
> > The thing I worry about in that style is object management. With a BroadcastReceiver or a ContentProvider I can follow the same lifecycle as is existing with the platform. If I move onto other classes like that, either the class will have to have an empty constructor and just get constructed by remote callbacks whenever it feels like, or we'll have to rely on an instance existing to be able to trigger callbacks. Both of these don't feel great and seem like they make it much harder to explain the situation in comparison to something simpler like a BroadcastReceiver where it follows the platform.
> 
> ack

Does this ensure that the content provider is an actual manifest component? Or can I define a content-provider just in code and use this for creating pending-intents (as the actual actions goes through a relay)?


Also is there support for processors in command-line tools like blaze and make?

Patch-set: 8
Reviewer: Gerrit User 1060398 <1060398@85c56323-6fa9-3386-8a01-6480fb634889>