Add the ability to allow for fragment addition instead of replacement
Add an extra NavOption to allow the user to specify whether the next destination Fragment should be added on top of or replace the previous Fragment. This is useful when the previous fragment is required to still be visible underneath the current one.

Replace is still the default behaviour, as it currently is.

Test: updated FragmentNavigatorTest
Change-Id: I2ee84d201a645c3d200bb962255b1266c0499e5e