Update patch set 30
Patch Set 30:

> Patch Set 30:
> 
> (2 comments)
> 
> Comparing the patches that returns null vs. non-null (patchset 30 vs.27), I think we should go with the one that return null. In my opinion the previous patch introduces more subtle behavior, i.e. for example the Destination class declares public objects ETA_NOT_SET and LOCATION_NOT_SET to prevent returning a null value. However, the app may still not expect to deal with an unknown value and throw an exception in a different way, or as it was mentioned in the patch, these objects can be modified by the app causing yet another problem,and dealing with this cause yet another inconsistent API that would create even more subtle behavior.
> 
> Returning null seems better, first, the API can be made more consistent and second and more important it allows the IDE and linter help the app developer address the unknown value issue in a sensible way. It is better to rely on tools in this case.
> 
> If the reviewers still insist, I think the best path is keeping the current null behavior and landing the CL and then adding a special note for API council to discuss it during review. API council reviews the APIs regularly so there should not be any concern on that.
> 
> There are other council members as reviewers in this CL. Alan or Jason should chime in if they have different opinions.

Thanks Mark, that is a pretty impressive document discussing both sides in good detail. I think in the end there is no clear winner. On the other hand, an app that receives an unexpected value from an API can still crash in a different way, even though it is not a null pointer exception. 
By returning a null, we at least take the advantage of automated tools indicating that an invalid value should be expected.

The more important is, in Android we should try not to change the behavior of APIs without an API bump. Apps that are targeting the previous API level should not see a change in behavior.

Patch-set: 30
Reviewer: Gerrit User 1010674 <1010674@85c56323-6fa9-3386-8a01-6480fb634889>