Add API to allow presenter to enforce header styling on the first row
Slices where the first row item is from a RowBuilder rather than a
HeaderBuilder don’t get the larger header styling treatment. We do this
by default because RowBuilder indicates that content is same level as
the rest of the slice whereas HeaderBuilder indicates that content is
of greater importance.

Multiple clients would like the ability to override this behavior for
their UX so let’s add the API.

Test: ./gradlew slice-view:connectedCheck
      manual - look at gallery and ride slice in slice browser, gallery
               gallery and rides first rows should be styled like a
               header
Bug: 112701769

Change-Id: I7488b098a297a30109ee90598c69a0891125f4bb