Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Does this break unbundled clones?
> 
> We don't have any branches that uses this anymore, it used to be used in ub-supportlib-master in AOSP, but we no longer support that one anyway and this does not merge to it.

I thought this was for external contributors so that they didn't need the entire checkout? I have this file locally so that I don't have to deal with changing my Android Studio SDK 6 times a day when I switch from a normal project to the support library.

Patch-set: 1