MCV2: Add/Remove custom playback speed
When an application sets a custom playback speed, check if the custom
value exists among the default playback speeds. If it doesn't
exist, add that value to the default speed list and update the
settings UI. When a default speed is selected again, remove the
custom value from the list and update the settings UI again.

Bug: 112821106
Test: Manually
Change-Id: I2da3089008e14025633ba5e455735328c079507e