Room FTS Support - Migration
Added fts entity migration validator that uses the sql create statement
for validation. This is necessary since fts options are not stored in
any public metadata table nor are available through any known pragma.

Created a FtsTableInfoValidatorWriter that will generate the code that
uses TableCreateInfo to get the create sql from sqlite_master so that it
can be compared and validated. TableInfoValidationWriter and
FtsTableInfoValidatorWriter now implement a common interface named
ValidationWriter.

Added a TypeAdapterFactory to SchemaBundle's GSON instance so that both
FtsEntityBundle and EntityBundle can be deserialized from the schema
json file.

Bug: 62356416
Test: ./gradlew room:room-compiler:test \
                room:integration-tests:testapp:cC

Change-Id: I34c4d0206ef75f47e1541f7c1f0543c0f2b1e853